Moved some tests to validation suite
